// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  EMPLOYER: 'EMPLOYER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const JobType = {
  FULLTIME: 'FULLTIME',
  PARTTIME: 'PARTTIME',
  CONTRACT: 'CONTRACT',
  INTERNSHIP: 'INTERNSHIP',
  REMOTE: 'REMOTE'
} as const

export type JobType = (typeof JobType)[keyof typeof JobType]


export const ApplicationStatus = {
  PENDING: 'PENDING',
  REVIEWING: 'REVIEWING',
  REJECTED: 'REJECTED',
  ACCEPTED: 'ACCEPTED'
} as const

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus]


export const NotificationType = {
  APPLY_JOB: 'APPLY_JOB',
  APPLICATION_STATUS: 'APPLICATION_STATUS',
  NEW_JOB: 'NEW_JOB'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]